refactor to always use sh and eliminate our own command parsing
