+++ webtau-feature-testing/examples/scenarios/cli/pathBasedScript.groovy
@@ -1,0 +1,17 @@
+package scenarios.cli
+
+import static com.twosigma.webtau.WebTauGroovyDsl.*
+
+scenario("cli run using path") {
+    cli.run('path-simple') {
+        output.should contain('version:')
+    }
+}
+
+scenario("cli run with args") {
+    cli.run('path-simple foo "hello world"') {
+        output.should contain('version:')
+        output.should contain('foo')
+        output.should contain('hello world')
+    }
+}
