+++ webtau-feature-testing/examples/scenarios/cli/simpleRun.groovy
@@ -7,17 +7,3 @@ scenario("simple cli run") {
         output.should contain('version:')
     }
 }
-
-scenario("cli run using path") {
-    cli.run('path-simple') {
-        output.should contain('version:')
-    }
-}
-
-scenario("cli run with args") {
-    cli.run('path-simple foo "hello world"') {
-        output.should contain('version:')
-        output.should contain('foo')
-        output.should contain('hello world')
-    }
-}
