+++ webtau-feature-testing/examples/scenarios/cli/webtau-with-path.cfg
@@ -1,0 +1,1 @@
+envPath = ['../scripts/more']
