--- webtau-feature-testing/examples/scenarios/cli/webtau.cfg
@@ -1,1 +1,0 @@
-envPath = ['../scripts/more']
