+++ webtau-feature-testing/test-expectations/scenarios/cli/pathBasedScript/run-details.json
@@ -1,0 +1,16 @@
+{
+  "scenarioDetails" : [ {
+    "scenario" : "cli run using path",
+    "shortContainerId" : "pathBasedScript.groovy",
+    "stepsSummary" : {
+      "numberOfSuccessful" : 2
+    }
+  }, {
+    "scenario" : "cli run with args",
+    "shortContainerId" : "pathBasedScript.groovy",
+    "stepsSummary" : {
+      "numberOfSuccessful" : 4
+    }
+  } ],
+  "exitCode" : 0
+}
